requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
opencv-python>=4.5.0
PyPDF2>=2.0.0
python-docx>=0.8.11
